Add GemPy tests and delete space functionality
Introduced GemPy-dependent tests for mesh uploads and data operations. Added a new `delete_space` function and its schema, enabling space deletion in the API. Updated and refactored test cases for improved organization and coverage.
