+++ liquid_earth_api/modules/rest_client/rest_interface.py
@@ -1,5 +1,5 @@
 from liquid_earth_api.modules.rest_client._utils import BASE_URL, handle_response
-from liquid_earth_api.data.schemas import AddNewSpacePostData, AddDataPostData
+from liquid_earth_api.data.schemas import AddNewSpacePostData, AddDataPostData, DeleteSpacePostData
 import requests
 from dataclasses import asdict
 
@@ -32,12 +32,24 @@ def get_available_projects(token: str) -> list:
 
 
 def post_create_space(add_new_space: AddNewSpacePostData, token: str) -> dict:
+    
+
     response = requests.post(
-        url=f"{BASE_URL}/AddNewSpace",
+        url=f"{BASE_URL}/AddNewSpace/v2",
         json=asdict(add_new_space),
         headers={"Authorization": f"{token}"}
     )
     return handle_response(response)
+
+
+def delete_space(delete_space_post_data: DeleteSpacePostData, token: str) -> dict:
+    response = requests.delete(
+        url=f"{BASE_URL}/DeleteSpace",
+        json=asdict(delete_space_post_data),
+        headers={"Authorization": f"{token}"}
+    )
+    return handle_response(response)
+
 
 def get_dev_tokens(user_id: str, login_token: str) -> dict:
     response = requests.get(
@@ -56,6 +68,7 @@ def generate_dev_token(user_id: str, token_name: str, login_token: str) -> dict:
         headers={"Authorization": f"{login_token}"}
     )
     return handle_response(response)
+
 
 def revoke_dev_token(user_id: str, token_id: str, login_token: str) -> dict:
     response = requests.post(
